requirements.txt
appdirs==1.4.2
boto3==1.4.4
botocore==1.5.18
click==6.7
docutils==0.13.1
jmespath==0.9.1
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
